Module intrusive_containers::linked_list

An intrusive double-linked list.

The 'LinkedList' allows elements to be inserted or removed from either end.

Structs

IntoIter

Iter
An iterator over references to the items of a LinkedList

IterMut
An iterator over mutable references to the items of a LinkedList

LinkedList
An intrusive doubly-linked list

Links

NodeImpl

Traits

Linkable
Link trait allowing a struct to be inserted into a LinkedList

Node
A trait that allows a struct to be inserted into a LinkedList

OwningPointer
A trait that allows insertion into a LinkedList. The trait is unsafe to implement due to the following constraints: 1) The deref functions must always return the same reference 2) The object cannot be moved while in the LinkedList 3) No references (mutable or otherwise) to the target can be used while list operations are ongoing